Add support for changing template style in iPad version. Also clean up a bunch of noisy warnings about Foundation classes returning NSInteger, which is not the same as int on iPad.

+++ src/Engine/EntireLayout.m
@@ -470,7 +470,7 @@ NSString *NormalizeDivisionString(NSString *inString) {
     }
     NSNumber *freightCarLength = fc.length;
     NSMutableSet *potentialLengthsToShow = [NSMutableSet set];
-    int carsToAdd = 3;
+    NSInteger carsToAdd = 3;
     if (freightCarLength) {
         [potentialLengthsToShow addObject: freightCarLength];
         carsToAdd = 2;
